Add -skipLookup flag to omit check for existing A record
This is a cludge around the initial case, where we haven't yet
added an A record for the host. Otherwise we get a 'host not found'
error, causing the script to exit.

I was tempted to put a bunch of logic in to handle the host not found
case, but tbh i'm more concerned that this would cause multiple api
calls when run continuously.

Instead, implement a 'first run, get the A record created' mode using
-skipLookup